Refactor Responses loop backend into cohesive private modules
